Extract ShowColumnsStatement to reduce duplicate logic for MySQL and doris
