+++ infra/binder/src/main/java/org/apache/shardingsphere/infra/binder/context/statement/SQLStatementContextFactory.java
@@ -71,6 +71,7 @@ import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.ExplainStat
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.FlushStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.KillStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.OptimizeTableStatement;
+import org.apache.shardingsphere.sql.parser.sql.common.statement.dal.ShowColumnsStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dcl.DCLStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dcl.GrantStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dcl.RevokeStatement;
@@ -102,7 +103,6 @@ import org.apache.shardingsphere.sql.parser.sql.common.statement.dml.DoStatement
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dml.InsertStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dml.SelectStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.dml.UpdateStatement;
-import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowColumnsStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowCreateTableStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowIndexStatement;
 import org.apache.shardingsphere.sql.parser.sql.dialect.statement.mysql.dal.MySQLShowTableStatusStatement;
@@ -267,8 +267,8 @@ public final class SQLStatementContextFactory {
         if (sqlStatement instanceof MySQLShowCreateTableStatement) {
             return new ShowCreateTableStatementContext((MySQLShowCreateTableStatement) sqlStatement);
         }
-        if (sqlStatement instanceof MySQLShowColumnsStatement) {
-            return new ShowColumnsStatementContext((MySQLShowColumnsStatement) sqlStatement);
+        if (sqlStatement instanceof ShowColumnsStatement) {
+            return new ShowColumnsStatementContext((ShowColumnsStatement) sqlStatement);
         }
         if (sqlStatement instanceof MySQLShowTablesStatement) {
             return new ShowTablesStatementContext((MySQLShowTablesStatement) sqlStatement);
